+++ src/main/webapp/teacher/createGame.html
@@ -187,6 +187,7 @@
   
     <script src="../scripts/main.js"></script>
     <script src="../scripts/auth.js"></script>
+    <script src="../scripts/createGameElements.js"></script>
     <script src="../scripts/createGame.js"></script>
   </body>
 </html>
